// src/gen/api/types/Friend.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { UserSummary } from './UserSummary.ts'

export type Friend = {
  /**
   * @type object
   */
  user: UserSummary
  /**
   * @description ISO 8601 timestamp
   * @type string, date-time
   */
  acceptedAt: string
}